+++ HISTORY.rst
@@ -5,6 +5,7 @@ History
 0.4.0 (not yet)
 ---------------
 
+* convert RGBA mode PNG images to RGB (to enable saving as JPG)
 * documentation for colorfix
 
 0.3.0 (2019-02-17)